backend/coding_contest/requirements.txt
appwrite==6.1.0
asgiref==3.8.1
CacheControl==0.14.1
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.0
Django==5.0.7
django-cors-headers==4.4.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.15.2
drf-yasg==1.21.8
firebase-admin==6.6.0
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
grpcio==1.68.1
grpcio-status==1.68.1
httplib2==0.22.0
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msgpack==1.1.0
oauth2client==4.1.3
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==24.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.1.2
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.0
uritemplate==4.1.1
urllib3==2.2.2
